requirements.txt
pandas>=1.5
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
streamlit>=1.20
